renamed TicketOrder-paths-final-amount.vue
